conditional rendering on progress page
